Charts refactoring to follow helm conventions: PNDA components to templates/, external requirements to charts/.

+++ scripts/package.sh
@@ -1,15 +1,28 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
-rm -f docs/*.tgz
-rm -f docs/*.yaml
+. scripts/common.sh
 
-for CHART in hdfs hbase opentsdb spark-standalone pnda
-do
+echo -e "${RS}${BLUE} Cleaning and preparing..."
+find . -iname "*.tgz" -type f | grep -v 'docs/' | xargs rm
 
-helm dep update $CHART
-helm package -d docs $CHART
-rm -rf $CHART/charts
+echo -e "${RS}${BLUE} Linting and checking...${RS}"
+scripts/lint.sh
+RET=$?
 
-done
+echo -e "${RS}${BLUE} Packaging charts... ${RS}"
+helm package pnda-helm-chart
+RET=$?
 
-helm repo index docs/
+if [[ "${RET}" == 0 ]]; then
+    helm repo index pnda-helm-chart
+
+    echo -e "Copying chart packages and index to docs ${RS}"
+    rm -f docs/*.tgz
+    rm -f docs/*.yaml
+    mv ./*.tgz docs/
+    mv ./pnda-helm-chart/index.yaml docs/
+
+    echo -e "${BLUE} Now commit and push charts and docs! ${RS}"
+else
+    echo -e "${RED}Tests failed, charts not packaged! ${RS}"
+fi
